Activity: Cast activity properties as integers where appropriate.
Props DJPaul, r-a-y.

See #6977.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@10853 cdf35c40-ae34-48e0-9cc9-0c9da1808c22

+++ src/bp-activity/bp-activity-screens.php
@@ -281,7 +281,7 @@ function bp_activity_screen_single_activity_permalink() {
 		} else {
 			$url = sprintf(
 				site_url( 'wp-login.php?redirect_to=%s' ),
-				urlencode( esc_url_raw( bp_activity_get_permalink( (int) bp_current_action() ) ) )
+				urlencode( esc_url_raw( bp_activity_get_permalink( bp_current_action() ) ) )
 			);
 		}
